The file is now read in memory and passed to openpyxl
Because openpyxl doesn't close the file when read with read-only, the file is now read with BytesIO and then closed.
ArgentaStatementParser can now be used in a 'with' statement.
convert.py can now import ofxstatement (ofxstatement-be-argenta wasn't installed next to ofxstatement).
convert.py renames its xlxs and ofx file to the accountid and last statement date.